vini-back/requirements.txt
fastapi==0.135.3
uvicorn[standard]==0.42.0
sqlmodel==0.0.22
python-dotenv==1.2.2
python-multipart==0.0.18
pydantic-settings>=2.7,<3
